Changes components → Network, a class with overridables.
While I originally wanted to keep the components as functions to be used beyond the scope of this project, the introduction of monitors precluded that. As such the move to this better organization changes little in terms of functionality; instead, within each component (Arrivals, Router, Job), would be interesting to try to decouple the monitoring from the main process logic, meeting original goal of having components be independent from the rest of the package.
